+++ src/emulator/hub.ts
@@ -33,8 +33,8 @@ export class EmulatorHub implements EmulatorInstance {
   static EMULATOR_HUB_ENV = "FIREBASE_EMULATOR_HUB";
   static CLI_VERSION = pkg.version;
   static PATH_EXPORT = "/_admin/export";
-  static PATH_DISABLE_FUNCTIONS = "/_admin/functions/disable";
-  static PATH_ENABLE_FUNCTIONS = "/_admin/functions/enable";
+  static PATH_DISABLE_FUNCTIONS = "/functions/disableBackgroundTriggers";
+  static PATH_ENABLE_FUNCTIONS = "/functions/enableBackgroundTriggers";
   static PATH_EMULATORS = "/emulators";
 
   /**