+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Updates the Google Cloud Run proxy API calls to `v1` (from `v1alpha1`) (#2695).
 - Release RTDB emulator v4.6.0: Get wire protocol with optional query.
+- Add endpoints to enable/disable background triggers in the Cloud Functions emulator.