+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-- Fixes issue where database export does not work if database is empty (#2634).
+- Updates the Google Cloud Run proxy API calls to `v1` (from `v1alpha1`) (#2695).
+- Release RTDB emulator v4.6.0: Get wire protocol with optional query.
 - Updates Cloud Functions for Firebase templates to better support function development.